+++ src/DDTrace/Configuration/AbstractConfiguration.php
@@ -49,6 +49,16 @@ abstract class AbstractConfiguration implements Registry
     /**
      * {@inheritdoc}
      *
+     * Returns a floating point value from the registry.
+     */
+    public function floatValue($key, $default, $min = null, $max = null)
+    {
+        return $this->registry->floatValue($key, $default, $min, $max);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
      * Allows users to access configuration properties by name instead of calling explicit methods.
      */
     public function inArray($key, $name)