+++ src/DDTrace/Configuration/EnvVariableRegistry.php
@@ -41,6 +41,34 @@ class EnvVariableRegistry implements Registry
     /**
      * {@inheritdoc}
      */
+    public function floatValue($key, $default, $min = null, $max = null)
+    {
+        if (!isset($this->registry[$key])) {
+            $value = getenv($this->convertKeyToEnvVariableName($key));
+            $value = trim(strtolower($value));
+            if (is_numeric($value)) {
+                $floatValue = (float)$value;
+            } else {
+                $floatValue = (float)$default;
+            }
+
+            if (null !== $min && $floatValue < $min) {
+                $floatValue = $min;
+            }
+
+            if (null !== $max && $floatValue > $max) {
+                $floatValue = $max;
+            }
+
+            $this->registry[$key] = $floatValue;
+        }
+
+        return $this->registry[$key];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function inArray($key, $name)
     {
         if (!isset($this->registry[$key])) {